mapping position support
Add mapping_order resource which manages the enable and position attributes for all mappings. All mappings in the remote server should be managed by this resource.

Mapping position is accomplished by creating a higher order resource which manages all the mapping positions. Mapping positions need to be managed in a single resource since changes in the position of one mapping cause a side effect where other mapping's positions are are updated.